int64 -> int32

+++ bfs.cc
@@ -4,7 +4,7 @@
 * BFS for optimal paths construction
 */
 
-void bfs(int maxelements, int MaxM, long *edges,  // matrix [maxelements, MaxM]
+void bfs(int maxelements, int MaxM, int *edges,  // matrix [maxelements, MaxM]
          int k, long *initial_vertex_ids,         // vector [n_queries]
          int m, long *gts,                        // vector [n_queries]
          int d1, int d2, long *distances,         // matrix [n_queries, maxelements]
