+++ bfs.h
@@ -15,7 +15,7 @@ struct Vertex{
     size_t min_path_length = 0;
 };
 
-void bfs(int maxelements, int MaxM, long *edges,  // matrix [maxelements, MaxM]
+void bfs(int maxelements, int MaxM, int *edges,  // matrix [maxelements, MaxM]
          int k, long *initial_vertex_ids,         // vector [n_queries]
          int m, long *gts,                        // vector [n_queries]
          int d1, int d2, long *distances,         // matrix [n_queries, maxelements]
